+++ Metal/Client/Client/StBodyL.cpp
@@ -25,8 +25,8 @@ void CStBodyL::handle(HDC hdc)
                 , (*iter_begin)->image
                 , -10 + (*obj)->GetInfo().fX - (*iter_begin)->image->GetWidth() / 2
                 , (*obj)->GetInfo().fY - (*iter_begin)->image->GetHeight() / 2
-                , (*iter_begin)->image->GetWidth() //(*iter_begin)->image->GetWidth()
-                , (*iter_begin)->image->GetHeight() //(*iter_begin)->image->GetHeight()
+                , FLOAT((*iter_begin)->image->GetWidth()) //(*iter_begin)->image->GetWidth()
+                , FLOAT((*iter_begin)->image->GetHeight()) //(*iter_begin)->image->GetHeight()
             );
         }
